requirements.txt
future
streamlit
pydub
whisper-openai